requirements.txt
biopython
plotly
pandas
scipy
streamlit>=0.83.0
watchdog